pyproject.toml
[project]
name = "celine-chatbot-api"
version = "0.3.1"
description = "FastAPI RAG chatbot using LlamaIndex + Qdrant + OpenAI"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "qdrant-client>=1.10.0",
  "llama-index>=0.11.0",
  "llama-index-vector-stores-qdrant>=0.3.0",
  "llama-index-embeddings-openai>=0.2.0",
  "openai>=1.40.0",
  "markitdown>=0.0.1",
  "python-jose[cryptography]>=3.3.0",
  "fsspec>=2024.6.0",
  "s3fs>=2024.6.0",
  "python-multipart>=0.0.9"
]

[tool.uv]
dev-dependencies = [
  "dumpster-cli>=1.4.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
